+++ dataflow/src/main/java/org/checkerframework/dataflow/expression/ArrayCreation.java
@@ -75,13 +75,13 @@ public class ArrayCreation extends JavaExpression {
   }
 
   @Override
-  public boolean isUnassignableByOtherCode() {
-    return false;
+  public boolean isAssignableByOtherCode() {
+    return true;
   }
 
   @Override
-  public boolean isUnmodifiableByOtherCode() {
-    return false;
+  public boolean isModifiableByOtherCode() {
+    return true;
   }
 
   @Override