+++ framework/src/main/java/org/checkerframework/framework/stub/AnnotationFileParser.java
@@ -555,7 +555,7 @@ public class AnnotationFileParser {
               // Find compile time constant fields, or values of an enum
               putAllNew(result, annosInType(element));
               importedConstants.addAll(getImportableMembers(element));
-              addEnclosingTypesToImportedTypes(element);
+              addEnclosedTypesToImportedTypes(element);
             }
 
           } else {
@@ -563,7 +563,7 @@ public class AnnotationFileParser {
             PackageElement element = findPackage(imported, importDecl);
             if (element != null) {
               putAllNew(result, annosInPackage(element));
-              addEnclosingTypesToImportedTypes(element);
+              addEnclosedTypesToImportedTypes(element);
             }
           }
         } else {
@@ -631,9 +631,9 @@ public class AnnotationFileParser {
    *
    * @param element an element for a type or package
    */
-  private void addEnclosingTypesToImportedTypes(Element element) {
+  private void addEnclosedTypesToImportedTypes(Element element) {
     for (Element enclosedEle : element.getEnclosedElements()) {
-      if (enclosedEle.getKind().isClass()) {
+      if (enclosedEle.getKind().isClass() || enclosedEle.getKind().isInterface()) {
         importedTypes.put(enclosedEle.getSimpleName().toString(), (TypeElement) enclosedEle);
       }
     }